refactor(agent-runtime): fold the Pi adapter into the Runtime directory
The target architecture puts everything Pi-specific in one directory, so
swapping the Runtime means deleting one directory and re-pointing one
composition line. Until now the Pi zone was two directories and the lint
rule had to name both.

Move `piApiAdapters.ts`, `piLanguageBinding.ts`, and `piModelResolver.ts`
with their tests into `agent/runtime/pi/`. Contents are unchanged; only
import paths move.

The Pi zone now inherits the Runtime directory's purity rule, which
`piModelResolver.ts` cannot satisfy: reading Provider and Model records
and materializing an Expo-backed fetch is exactly its job. Lint scopes
that exemption to that one file — the same boundary the PiRuntime
conformance harness already draws with its `sourceFiles` list, so the two
enforcement mechanisms now agree on which files must stay pure.
